Copy emoji symbol, or with modifier, emoji code.
This changes the output of the main script to always output the emoji code,
and adjusts the Alfred workflow to take that emoji code and turn it into
the UTF-8 emoji symbol (using emoji_to_symbol.rb) which then gets copied
to the clipboard. If the `alt` key is pressed, then the emoji code will
be copied to the clipboard instead.

Fixes #17